requirements.txt
pandas
numpy
matplotlib
jupyter
pyyaml
scikit-learn
